refactor the proxy middleware
