+++ include/http/middleware/index.js
@@ -21,11 +21,11 @@ const next = fn => (req, res, next) => Promise.resolve(fn).then(fn => fn(req, re
 module.exports = function(pb) {
     const { startTime, endTime } = require('./timing')(pb)
     const { openSession, closeSession, writeSessionCookie } = require('./session')(pb)
-    const { urlParse, checkPublicRoute, checkModuleRoute, systemSetupCheck, parseRequestBody } = require('./system')(pb)
+    const { urlParse, callProxy, checkPublicRoute, checkModuleRoute, systemSetupCheck, parseRequestBody } = require('./system')(pb)
     const { deriveSite, deriveActiveTheme, deriveRoute, inactiveAccessCheck } = require('./routing')(pb)
     const { requiresAuthenticationCheck, authorizationCheck, ipFilterCheck } = require('./auth')(pb)
     const { localizedRouteCheck, initializeLocalization } = require('./localization')(pb)
-    const { instantiateController, initializeController, render, writeResponse, prefix, callProxy } = require('./controller')(pb)
+    const { instantiateController, initializeController, render, writeResponse, prefix } = require('./controller')(pb)
 
     const stack = [
         startTime,
