+++ include/http/middleware/system.js
@@ -3,7 +3,7 @@ const util = require('util');
 const url = require('url');
 const HttpStatus = require('http-status-codes');
 const ErrorUtils = require('../../error/error_utils');
-
+const request = require('request-promise');
 
 const publicRoutes = ['/js/', '/css/', '/fonts/', '/img/', '/localization/', '/favicon.ico', '/docs/'];
 const modulePattern = /^\/node_modules\/(.*)/
@@ -23,6 +23,59 @@ module.exports = pb => ({
             req.handler.url = url.parse(req.url, true);
         }
     },
+
+    async callProxy(req, res) {
+        if (req.url.startsWith('proxy') || req.url.startsWith('/proxy')) {
+            const proxyDomain = `https://tn-consumer-poc.herokuapp.com`;
+            const url = req.url.replace('/proxy', proxyDomain);
+
+            try {
+                const response = await request({
+                    url: url,
+                    method: 'GET',
+                    headers: {
+                        "Accept": '*/*;q=0.5, text/javascript, application/javascript, application/ecmascript, application/x-ecmascript',
+                        "Referer": 'http://premium.lvh.me:8080/searchv2',
+                        'Origin': 'http://premium.lvh.me:8080',
+                        'User-Agent': 'Mozilla/ 5.0(Macintosh; Intel Mac OS X 10_13_6) AppleWebKit / 537.36(KHTML, like Gecko) Chrome / 75.0.3770.90 Safari / 537.36',
+                        'DNT': '1',
+                    },
+                    resolveWithFullResponse: true
+                });
+
+                let conetnt = response.body;
+                const contentType = response.headers['content-type'];
+
+                // if (contentType.includes('javascript')) {
+                //     conetnt = conetnt.replace(/(\$\.ajax[\s\S]*url[^\/]*)(\/)/, function (match, p1, p2) {
+                //         return `${p1}/proxy/`;
+                //     })
+                // }
+
+                if (contentType.includes('javascript')) {
+                    conetnt = conetnt.replace(`url: host + '/jobs.js'`, function (match, p1, p2) {
+                        return `url: '/proxy/jobs.js'`;
+                    })
+                }
+
+                conetnt = conetnt.replace(/(<[link].*[href\src]\=['|"])(\/)/g, function (match, p1, p2) {
+                    return `${p1}/proxy/`;
+                });
+                conetnt = conetnt.replace(/(<[script|a].*[href\src]\=['|"])(\/)/g, function (match, p1, p2) {
+                    return `${p1}/proxy/`;
+                });
+
+                req.handler.writeResponse({
+                    content: conetnt, content_type: contentType
+                });
+
+                req.router.continueAfter('writeResponse');
+            } catch (ex) {
+                console.log(ex);
+            }
+        }
+    },
+
     checkPublicRoute: async (req, res) => {
         let pathname = req.handler.url.pathname;
